requirements.txt
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
ffmpeg-python==0.2.0
huggingface-hub==0.29.1
ipython==9.0.0
librosa==0.10.2.post1
numba==0.61.0
numpy==2.1.3
openai==1.65.1
PyAudio==0.2.13
pydub==0.25.1
soundfile==0.13.1
tokenizers==0.21.0
torch==2.6.0
transformers==4.49.0
google-cloud-translate
